Prefer jakarta apis over older javax apis.
Update to use the latest jakarta apis in all dropwizard modules instead
of a mix of jakarta and javax.* ones (which ship the same classes). This
should make the APIs on the classpath consistent when creating a shaded
application jar and also avoid warnings about duplicate classes during
shading.
