+++ dropwizard-dependencies/pom.xml
@@ -18,6 +18,7 @@
 
     <properties>
         <!-- External dependencies -->
+        <activation-api.version>1.2.1</activation-api.version>
         <argparse4j.version>0.8.1</argparse4j.version>
         <byte-buddy.version>1.10.6</byte-buddy.version>
         <caffeine.version>2.8.0</caffeine.version>
@@ -35,9 +36,9 @@
         <jackson.version>2.10.1</jackson.version>
         <jakarta.annotation-api.version>1.3.5</jakarta.annotation-api.version>
         <jakarta.el.version>3.0.3</jakarta.el.version>
+        <jakarta.servlet-api.version>4.0.3</jakarta.servlet-api.version>
         <javassist.version>3.26.0-GA</javassist.version>
-        <javax.servlet-api.version>4.0.1</javax.servlet-api.version>
-        <jaxb-api.version>2.3.1</jaxb-api.version>
+        <jaxb-api.version>2.3.2</jaxb-api.version>
         <jcip-annotations.version>1.0</jcip-annotations.version>
         <jdbi3.version>3.12.0</jdbi3.version>
         <jersey.version>2.29.1</jersey.version>
@@ -53,7 +54,6 @@
         <mustache-compiler.version>0.9.6</mustache-compiler.version>
         <nullaway.version>0.7.8</nullaway.version>
         <objenesis.version>3.1</objenesis.version>
-        <servlet.version>3.0.0.v201112011016</servlet.version>
         <slf4j.version>1.7.30</slf4j.version>
         <tomcat-jdbc.version>9.0.30</tomcat-jdbc.version>
         <usertype.core.version>7.0.0.CR1</usertype.core.version>
@@ -155,26 +155,26 @@
                 <version>${jakarta.el.version}</version>
             </dependency>
             <dependency>
+                <groupId>jakarta.activation</groupId>
+                <artifactId>jakarta.activation-api</artifactId>
+                <version>${activation-api.version}</version>
+                <scope>runtime</scope>
+            </dependency>
+            <dependency>
                 <groupId>jakarta.annotation</groupId>
                 <artifactId>jakarta.annotation-api</artifactId>
                 <version>${jakarta.annotation-api.version}</version>
             </dependency>
             <dependency>
-                <groupId>javax.servlet</groupId>
-                <artifactId>javax.servlet-api</artifactId>
-                <version>${javax.servlet-api.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>javax.xml.bind</groupId>
-                <artifactId>jaxb-api</artifactId>
+                <groupId>jakarta.xml.bind</groupId>
+                <artifactId>jakarta.xml.bind-api</artifactId>
                 <version>${jaxb-api.version}</version>
                 <scope>runtime</scope>
             </dependency>
             <dependency>
-                <groupId>javax.activation</groupId>
-                <artifactId>javax.activation-api</artifactId>
-                <version>1.2.0</version>
-                <scope>runtime</scope>
+                <groupId>jakarta.servlet</groupId>
+                <artifactId>jakarta.servlet-api</artifactId>
+                <version>${jakarta.servlet-api.version}</version>
             </dependency>
             <dependency>
                 <groupId>org.apache.httpcomponents</groupId>
